Bug 1197673 - Use float type for SetAudioOutputVolume. r=eitan

+++ dom/media/webspeech/synth/ipc/SpeechSynthesisParent.h
@@ -70,7 +70,7 @@ protected:
 
   virtual bool RecvForceEnd() override;
 
-  virtual bool RecvSetAudioOutputVolume(const uint32_t& aVolume) override;
+  virtual bool RecvSetAudioOutputVolume(const float& aVolume) override;
 };
 
 class SpeechTaskParent : public nsSpeechTask
